remove POSIXFileBasedLock for now
